requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
deepgram-sdk==3.7.6
deprecation==2.1.0
fastapi==0.115.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==3.0.2
marshmallow==3.23.1
minio==7.2.10
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
propcache==0.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==14.0
yarl==1.17.1
